refactor: move Lang type to src/i18n/types.ts
Lang is a site-wide i18n primitive, not a CV concern. All 22 importers
updated to import from i18n/types. data/cv.ts re-exports Lang for
backward compat with the few files that also import CV-specific types.
