+++ src/components/blog/BlogPage.astro
@@ -6,7 +6,7 @@ import DotField from '../atoms/DotField.astro';
 import Button from '../atoms/Button.astro';
 import { sortPostsByDate, filterDrafts } from '../../lib/blog';
 import { t, getBlogSlugFromId, getLocalePath } from '../../i18n/utils';
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 
 interface Props {
   lang: Lang;
