+++ src/components/blog/CodeBlockEnhancer.astro
@@ -15,7 +15,7 @@
 
 import CopyButton from '../atoms/CopyButton.astro';
 import { t } from '../../i18n/utils';
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 
 interface Props {
   lang: Lang;
