+++ src/components/case-study/SelfPage.astro
@@ -2,7 +2,7 @@
 import Layout from '../../layouts/Layout.astro';
 import Button from '../atoms/Button.astro';
 import { t } from '../../i18n/utils';
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import { fetchBuildStats } from '../../lib/build-stats';
 import StatCard from '../home/StatCard.astro';
 import SectionHeader from '../atoms/SectionHeader.astro';
