+++ src/components/cv/CvPage.astro
@@ -10,7 +10,7 @@ import CvPdfDownload from './CvPdfDownload.astro';
 import CvScrollReveal from './CvScrollReveal.astro';
 import DotField from '../atoms/DotField.astro';
 import { getCvData, getPdfPath, getJsonPath } from '../../data/cv';
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 
 interface Props {
   lang: Lang;
