+++ src/components/cv/CvPrintLayout.astro
@@ -12,7 +12,8 @@
  * - Zero decorative elements: no dots, borders, shadows, icons
  * - Skills rendered as inline comma-separated lists to save vertical space
  */
-import type { CvData, Lang } from '../../data/cv';
+import type { CvData } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 
 interface Props {
   data: CvData;
