+++ src/components/home/ContactSection.astro
@@ -10,7 +10,7 @@
  * (id="contact").
  */
 import { Icon } from 'astro-icon/components';
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import { t } from '../../i18n/utils';
 import Button from '../atoms/Button.astro';
 import SectionHeader from '../atoms/SectionHeader.astro';
