+++ src/components/home/HeroParallax.astro
@@ -22,7 +22,7 @@
  * Dot field positions are generated at build time for stability across renders.
  */
 
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import { t } from '../../i18n/utils';
 
 interface Props {
