+++ src/components/home/HomeContent.astro
@@ -7,7 +7,7 @@
  * variants, and the Layout/Nav behave normally over the top. Individual
  * sections handle their own animations, parallax and IntersectionObservers.
  */
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import HeroParallax from './HeroParallax.astro';
 import WhatIDo from './WhatIDo.astro';
 import StatsCounter from './StatsCounter.astro';
