+++ src/components/home/LatestPosts.astro
@@ -12,7 +12,7 @@ import PostCard from '../molecules/PostCard.astro';
 import SectionHeader from '../atoms/SectionHeader.astro';
 import { sortPostsByDate, filterDrafts } from '../../lib/blog';
 import { t } from '../../i18n/utils';
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 
 interface Props {
   lang: Lang;
