+++ src/components/home/Portfolio.astro
@@ -1,5 +1,5 @@
 ---
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import { t } from '../../i18n/utils';
 import SectionHeader from '../atoms/SectionHeader.astro';
 import EyebrowTag from '../atoms/EyebrowTag.astro';
