+++ src/components/home/StatsCounter.astro
@@ -9,7 +9,7 @@
  * Single IntersectionObserver, one rAF loop per counter, cleans itself up.
  * Respects prefers-reduced-motion (shows final value immediately).
  */
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import { t } from '../../i18n/utils';
 import StatCard from './StatCard.astro';
 import SectionHeader from '../atoms/SectionHeader.astro';
