+++ src/components/home/Talks.astro
@@ -1,5 +1,5 @@
 ---
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import { t } from '../../i18n/utils';
 import { talks } from '../../data/talks';
 import { formatDate, formatDateISO } from '../../lib/date';
