+++ src/components/home/TechStack.astro
@@ -4,7 +4,7 @@
  * Hydrated with client:visible so the interactive JS only loads when the
  * section actually enters the viewport.
  */
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import { t } from '../../i18n/utils';
 import TechConstellation from './TechConstellation.tsx';
 import SectionHeader from '../atoms/SectionHeader.astro';
