+++ src/components/home/Testimonials.astro
@@ -1,6 +1,6 @@
 ---
 import { t } from '../../i18n/utils';
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import TestimonialCard from './TestimonialCard.astro';
 import SectionHeader from '../atoms/SectionHeader.astro';
