+++ src/components/home/WhatIDo.astro
@@ -9,7 +9,7 @@
  * Uses IntersectionObserver + CSS custom properties (--stagger-delay) to
  * avoid JS animation loops. Respects prefers-reduced-motion.
  */
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import { t } from '../../i18n/utils';
 import ServiceCard from './ServiceCard.astro';
 import SectionHeader from '../atoms/SectionHeader.astro';
