+++ src/components/katas/KataCard.astro
@@ -1,7 +1,7 @@
 ---
 import Tag from '../atoms/Tag.astro';
 import type { Kata } from '../../data/katas';
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import { t } from '../../i18n/utils';
 
 interface Props {
