+++ src/components/katas/KatasPage.astro
@@ -4,7 +4,7 @@ import KatasScrollReveal from './KatasScrollReveal.astro';
 import DotField from '../atoms/DotField.astro';
 import Button from '../atoms/Button.astro';
 import { sortedKatas } from '../../data/katas';
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 import { getLocalePath, t } from '../../i18n/utils';
 
 interface Props {
