+++ src/components/legal/LegalPageLayout.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../../layouts/Layout.astro';
-import type { Lang } from '../../data/cv';
+import type { Lang } from '../../i18n/types';
 
 interface Props {
   title: string;
