+++ src/i18n/types.ts
@@ -1,0 +1,1 @@
+export type Lang = 'es' | 'en';
