+++ src/layouts/Layout.astro
@@ -6,7 +6,7 @@ import CookieConsent from '../components/analytics/CookieConsent.astro';
 import Nav from "../components/nav/Nav.astro";
 import Footer from "../components/footer/Footer.astro";
 import { getLangFromUrl, getAlternateUrl, t } from '../i18n/utils';
-import type { Lang } from '../data/cv';
+import type { Lang } from '../i18n/types';
 
 interface Props {
 	title: string;
